requirements.txt
monzo-api==0.2.1
black